重构项目结构，添加中英文readme
