+++ README.md
@@ -1,14 +1,47 @@
-开发中
+# VRChat 触摸反馈系统
 
-能用，但是bug还很多
+[English](./README_EN.md)
 
-使用郊狼来模拟被触摸的感觉，说人话就是可以通过触摸你的avatar的特定部位（指触发器所在的地方）来生成电击波形。
+通过 VRChat 的触摸交互实现真实的触觉反馈体验。
 
-支持多个郊狼设备同时连接到websocket上来实现通感玩法。
+## 功能特点
 
-设备需求是郊狼3.0
+- 支持通过触摸 Avatar 特定部位触发触觉反馈
+- 支持多个郊狼设备同时连接，实现多人互动
+- 完全兼容 MA 预制件
+- 可自定义触发区域位置
 
-把耳朵摸摸这个预制件塞到你的avatar里就可以，预制件支持MA，我这个耳朵摸摸是根据lime的耳朵做的，你也可以把触发器放到其他位置上，只要保持变量名一致就好.
+## 系统要求
 
-连接成功之后按A通道的圆圈按钮启动接收
+- 郊狼 3.0 设备
+- VRChat 账号
+- 支持 MA 的 Avatar
+
+## 安装说明
+
+1. 下载最新版本的预制件
+2. 将预制件导入到你的 Avatar 项目中
+3. 将触发器组件放置到需要触觉反馈的位置
+4. 确保所有变量名与预制件保持一致
+
+## 使用方法
+
+1. 启动 VRChat 并进入游戏
+2. 连接郊狼设备到 WebSocket
+3. 按下设备 A 通道的圆形按钮开始接收信号
+4. 当其他玩家触摸到触发区域时，将会产生相应的触觉反馈
+
+## 注意事项
+
+- 当前版本仍在开发中，可能存在一些已知问题
+- 建议在使用前检查设备连接状态
+- 如遇到问题请提交 Issue
+
+## 贡献指南
+
+欢迎提交 Pull Request 或 Issue 来帮助改进项目。
+
+## 许可证
+
+[许可证类型]
