+++ README_EN.md
@@ -1,0 +1,44 @@
+# VRChat Haptic Feedback System
+
+A haptic feedback system that provides real touch sensations through VRChat interactions.
+
+## Features
+
+- Haptic feedback triggered by touching specific areas of your Avatar
+- Support for multiple Coyote devices connected simultaneously
+- Full compatibility with MA prefabs
+- Customizable trigger zones
+
+## Requirements
+
+- Coyote 3.0 device
+- VRChat account
+- MA-compatible Avatar
+
+## Installation
+
+1. Download the latest version of the prefab
+2. Import the prefab into your Avatar project
+3. Place trigger components at desired feedback locations
+4. Ensure all variable names match the prefab requirements
+
+## Usage
+
+1. Launch VRChat and enter the game
+2. Connect your Coyote device to WebSocket
+3. Press the circular button on channel A to start receiving signals
+4. When other players touch the trigger zones, haptic feedback will be generated
+
+## Notes
+
+- Current version is under development and may contain known issues
+- Please check device connection status before use
+- Submit issues if you encounter any problems
+
+## Contributing
+
+Contributions via Pull Requests or Issues are welcome.
+
+## License
+
+[License Type]
